refactor: errorcodee 패키지 수정
