+++ blue-sol-backend/src/main/java/com/solif/backend/domain/auth/controller/AuthController.java
@@ -1,7 +1,8 @@
 package com.solif.backend.domain.auth.controller;
 
+import com.solif.backend.domain.auth.code.AuthSuccessCode;
 import com.solif.backend.domain.auth.dto.*;
-import com.solif.backend.domain.auth.exception.AuthErrorCode;
+import com.solif.backend.domain.auth.code.AuthErrorCode;
 import com.solif.backend.domain.auth.service.AuthService;
 import com.solif.backend.global.common.exception.CustomException;
 import com.solif.backend.global.common.response.ResponseFactory;
