+++ blue-sol-backend/src/main/java/com/solif/backend/domain/council/service/CouncilMemberService.java
@@ -1,6 +1,6 @@
 package com.solif.backend.domain.council.service;
 
-import com.solif.backend.domain.auth.exception.AuthErrorCode;
+import com.solif.backend.domain.auth.code.AuthErrorCode;
 import com.solif.backend.domain.council.code.CouncilErrorCode;
 import com.solif.backend.domain.council.dto.AddMemberResponse;
 import com.solif.backend.domain.council.dto.CouncilMemberResponse;
