+++ blue-sol-backend/src/main/java/com/solif/backend/domain/goal/code/GoalErrorCode.java
@@ -1,4 +1,4 @@
-package com.solif.backend.domain.goal.exception;
+package com.solif.backend.domain.goal.code;
 
 import com.solif.backend.global.common.exception.ErrorCode;
 import lombok.Getter;
