+++ blue-sol-backend/src/main/java/com/solif/backend/domain/network/code/NetworkSuccessCode.java
@@ -1,4 +1,4 @@
-package com.solif.backend.domain.network.dto;
+package com.solif.backend.domain.network.code;
 
 import com.solif.backend.global.common.response.SuccessCode;
 import lombok.Getter;
