+++ blue-sol-backend/src/main/java/com/solif/backend/global/jwt/JwtAuthenticationFilter.java
@@ -1,6 +1,6 @@
 package com.solif.backend.global.jwt;
 
-import com.solif.backend.domain.auth.exception.AuthErrorCode;
+import com.solif.backend.domain.auth.code.AuthErrorCode;
 import com.solif.backend.global.common.exception.CustomException;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
