+++ blue-sol-backend/src/main/java/com/solif/backend/global/s3/S3Service.java
@@ -1,6 +1,6 @@
 package com.solif.backend.global.s3;
 
-import com.solif.backend.domain.file.exception.FileException;
+import com.solif.backend.domain.file.code.FileException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -16,7 +16,7 @@ import software.amazon.awssdk.services.s3.model.S3Object;
 
 import java.util.List;
 
-import static com.solif.backend.domain.file.exception.FileException.FileErrorCode.FILE_UPLOAD_FAILED;
+import static com.solif.backend.domain.file.code.FileException.FileErrorCode.FILE_UPLOAD_FAILED;
 
 @Slf4j
 @Service
